requirements.txt
numpy==1.15.1
numpy==1.19.2
numpy==1.21.5
numpy==1.22.3
numpy==1.22.0
numpy==1.23.0
numpy==1.24.3
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
opencv_python==4.6.0.66
Pillow==8.0.1
Pillow==9.0.1
Pillow==9.2.0
Pillow==9.4.0
Pillow==9.5.0
Pillow==9.5.0
tensorflow==2.12.0
tensorflow==2.3.0
tensorflow==2.9.1
tensorflow_intel==2.12.0
